requirements.txt
Flask
OpenAI-whisper
matplotlib

